tests: Fix size_t vs int issues
size_t can be any unsigned type but is never a signed integer.

+++ tests/driver_sx127x/main.c
@@ -230,8 +230,8 @@ int send_cmd(int argc, char **argv)
         return -1;
     }
 
-    printf("sending \"%s\" payload (%d bytes)\n",
-           argv[1], strlen(argv[1]) + 1);
+    printf("sending \"%s\" payload (%u bytes)\n",
+           argv[1], (unsigned)strlen(argv[1]) + 1);
 
     iolist_t iolist = {
         .iol_base = argv[1],
